@@ -1,11 +1,12 @@
 package io.protostuff.compiler.cli;
 
-import com.google.common.base.Joiner;
-import com.google.common.base.MoreObjects;
 import com.google.common.collect.ImmutableMap;
+import io.protostuff.compiler.model.ImmutableModuleConfiguration;
 import io.protostuff.compiler.model.ModuleConfiguration;
+import io.protostuff.compiler.parser.ParserException;
 import io.protostuff.generator.CompilerModule;
 import io.protostuff.generator.CompilerRegistry;
+import io.protostuff.generator.GeneratorException;
 import io.protostuff.generator.ProtostuffCompiler;
 import org.apache.commons.cli.*;
 import org.apache.logging.log4j.Level;
@@ -16,8 +17,12 @@ import org.apache.logging.log4j.core.config.LoggerConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.util.*;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -81,7 +86,7 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
                 .longOpt(GENERATOR)
                 .argName("name")
                 .numberOfArgs(1)
-                .desc("Specify compiler: " + Joiner.on('|').join(registry.availableCompilers()))
+                .desc("Specify compiler: " + String.join("|", registry.availableCompilers()))
                 .build());
         options.addOption(Option.builder("t")
                 .longOpt(TEMPLATE)
@@ -96,8 +101,8 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
                 .desc("[st4] Specify full class name of an extensions provider for st4 compiler")
                 .build());
         CommandLineParser parser = new DefaultParser();
-        ModuleConfiguration configuration = new ModuleConfiguration();
-        configuration.setName("main");
+        ImmutableModuleConfiguration.Builder builder = ImmutableModuleConfiguration.builder();
+        builder.name("main");
         try {
             CommandLine cmd = parser.parse(options, args);
             if (cmd.hasOption(HELP)) {
@@ -115,40 +120,38 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
             }
             if (cmd.hasOption(GENERATOR)) {
                 String generator = cmd.getOptionValue(GENERATOR);
-                configuration.setGenerator(generator);
+                builder.generator(generator);
             } else {
                 LOGGER.error("Generator is not set.");
                 return;
             }
             if (cmd.hasOption(OUTPUT)) {
                 String output = cmd.getOptionValue(OUTPUT);
-                configuration.setOutput(output);
+                builder.output(output);
             } else {
                 LOGGER.error("Output directory is not set.");
                 return;
             }
-            Map<String, Object> moduleOptions = new HashMap<String, Object>();
             if (cmd.hasOption(TEMPLATE)) {
                 List<String> templates = Collections.singletonList(cmd.getOptionValue(TEMPLATE));
-                moduleOptions.put(CompilerModule.TEMPLATES_OPTION, templates);
+                builder.putOptions(CompilerModule.TEMPLATES_OPTION, templates);
             }
             if (cmd.hasOption(EXTENSIONS)) {
-                moduleOptions.put(CompilerModule.EXTENSIONS_OPTION, cmd.getOptionValue(EXTENSIONS));
+                builder.putOptions(CompilerModule.EXTENSIONS_OPTION, cmd.getOptionValue(EXTENSIONS));
             }
-            configuration.setOptions(moduleOptions);
-            List<File> includePaths = new ArrayList<File>();
+            List<Path> includePaths = new ArrayList<>();
             if (cmd.hasOption(PROTO_PATH)) {
                 String[] paths = cmd.getOptionValues(PROTO_PATH);
                 for (String path : paths) {
-                    includePaths.add(new File(path));
+                    includePaths.add(Paths.get(path));
                 }
             }
             if (includePaths.isEmpty()) {
-                includePaths.add(new File("."));
+                includePaths.add(Paths.get("."));
             }
-            configuration.setIncludePaths(includePaths);
+            builder.includePaths(includePaths);
             List<String> protoFiles = cmd.getArgList();
-            configuration.setProtoFiles(protoFiles);
+            builder.protoFiles(protoFiles);
         } catch (ParseException e) {
             if (LOGGER.isDebugEnabled()) {
                 LOGGER.error("Could not parse command", e);
@@ -160,6 +163,8 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
 
         LOGGER.info("Version={}", __VERSION);
 
+        ModuleConfiguration configuration = builder.build();
+
         if (configuration.getProtoFiles().isEmpty()) {
             LOGGER.error("Missing input file.");
             return;
@@ -170,7 +175,7 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
         }
         try {
             compile(configuration);
-        } catch (Exception e) {
+        } catch (GeneratorException | ParserException e) {
             if (LOGGER.isDebugEnabled()) {
                 LOGGER.error("Compilation error", e);
             } else {
@@ -183,7 +188,7 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
         // automatically generate the help statement
         HelpFormatter formatter = new HelpFormatter();
         formatter.setWidth(79);
-        final Map<String, Integer> order = ImmutableMap.<String, Integer>builder()
+        Map<String, Integer> order = ImmutableMap.<String, Integer>builder()
                 .put(HELP, 1)
                 .put(PROTO_PATH, 2)
                 .put(GENERATOR, 3)
@@ -192,18 +197,9 @@ public class ProtostuffCompilerCLI extends ProtostuffCompiler {
                 .put(EXTENSIONS, 6)
                 .put(DEBUG, 100)
                 .build();
-        formatter.setOptionComparator(new Comparator<Option>() {
-            @Override
-            public int compare(Option o1, Option o2) {
-                return compare(
-                        MoreObjects.firstNonNull(order.get(o1.getLongOpt()), 99),
-                        MoreObjects.firstNonNull(order.get(o2.getLongOpt()), 99));
-            }
-
-            private int compare(int x, int y) {
-                return (x < y) ? -1 : ((x == y) ? 0 : 1);
-            }
-        });
+        formatter.setOptionComparator((o1, o2) -> Integer.compare(
+                order.getOrDefault(o1.getLongOpt(), 99),
+                order.getOrDefault(o2.getLongOpt(), 99)));
         formatter.printHelp("protostuff-compiler [options] proto_files", options);
     }
 
