@@ -1,5 +1,7 @@
 package io.protostuff.generator;
 
+import java.util.function.Function;
+
 /**
  * @author Kostiantyn Shchepanovskyi
  */
@@ -9,5 +11,5 @@ public interface PropertyProvider<ObjectT> {
 
     Object getProperty(ObjectT object, String propertyName);
 
-    void register(String property, ComputableProperty<ObjectT, ?> function);
+    void register(String property, Function<ObjectT, ?> function);
 }
