@@ -2,13 +2,14 @@ package io.protostuff.generator;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.function.Function;
 
 /**
  * @author Kostiantyn Shchepanovskyi
  */
 public class SimplePropertyProvider<ObjectT> implements PropertyProvider<ObjectT> {
 
-    private final Map<String, ComputableProperty<ObjectT, ?>> propertyProviders = new HashMap<String, ComputableProperty<ObjectT, ?>>();
+    private final Map<String, Function<ObjectT, ?>> propertyProviders = new HashMap<>();
 
     @Override
     public boolean hasProperty(String propertyName) {
@@ -17,15 +18,15 @@ public class SimplePropertyProvider<ObjectT> implements PropertyProvider<ObjectT
 
     @Override
     public Object getProperty(ObjectT object, String propertyName) {
-        ComputableProperty<ObjectT, ?> provider = propertyProviders.get(propertyName);
+        Function<ObjectT, ?> provider = propertyProviders.get(propertyName);
         if (provider == null) {
             throw new IllegalArgumentException(propertyName);
         }
-        return provider.compute(object);
+        return provider.apply(object);
     }
 
     @Override
-    public void register(String property, ComputableProperty<ObjectT, ?> function) {
+    public void register(String property, Function<ObjectT, ?> function) {
         propertyProviders.put(property, function);
     }
 
