@@ -1,61 +1,32 @@
 package io.protostuff.generator.html.json.enumeration;
 
-import io.protostuff.generator.html.json.index.NodeType;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+import org.immutables.value.Value;
+
+import java.util.List;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.List;
+
+import io.protostuff.generator.html.json.index.NodeType;
 
 /**
  * @author Kostiantyn Shchepanovskyi
  */
-public class EnumDescriptor {
+@Value.Immutable
+@JsonSerialize(as = ImmutableEnumDescriptor.class)
+@JsonDeserialize(as = ImmutableEnumDescriptor.class)
+public interface EnumDescriptor {
 
-    private String name;
-    private NodeType type;
-    private String canonicalName;
-    @Nullable
-    private String description;
-    private List<EnumConstant> constants = new ArrayList<EnumConstant>();
+    String name();
 
-    public String getName() {
-        return name;
-    }
+    NodeType type();
 
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public NodeType getType() {
-        return type;
-    }
-
-    public void setType(NodeType type) {
-        this.type = type;
-    }
-
-    public String getCanonicalName() {
-        return canonicalName;
-    }
-
-    public void setCanonicalName(String canonicalName) {
-        this.canonicalName = canonicalName;
-    }
+    String canonicalName();
 
     @Nullable
-    public String getDescription() {
-        return description;
-    }
+    String description();
 
-    public void setDescription(@Nullable String description) {
-        this.description = description;
-    }
-
-    public List<EnumConstant> getConstants() {
-        return constants;
-    }
-
-    public void setConstants(List<EnumConstant> constants) {
-        this.constants = constants;
-    }
+    List<EnumConstant> constants();
 }
