@@ -1,12 +1,12 @@
 package io.protostuff.generator.html.json.proto;
 
+import javax.inject.Inject;
+
 import io.protostuff.compiler.model.Module;
 import io.protostuff.compiler.model.Proto;
 import io.protostuff.generator.OutputStreamFactory;
 import io.protostuff.generator.html.json.AbstractJsonGenerator;
 import io.protostuff.generator.html.json.index.NodeType;
-
-import javax.inject.Inject;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -20,19 +20,19 @@ public class JsonProtoGenerator extends AbstractJsonGenerator {
 
     @Override
     public void compile(Module module) {
-        for (Proto proto : module.getProtos()) {
-            process(module, proto);
-        }
+        module.getProtos().stream()
+                .forEach(proto -> process(module, proto));
     }
 
     private void process(Module module, Proto proto) {
         String name = proto.getCanonicalName();
-        ProtoDescriptor descriptor = new ProtoDescriptor();
-        descriptor.setName(proto.getName());
-        descriptor.setType(NodeType.PROTO);
-        descriptor.setCanonicalName(proto.getCanonicalName());
-        descriptor.setFilename(proto.getFilename());
-        descriptor.setDescription(proto.getComments());
+        ImmutableProtoDescriptor descriptor = ImmutableProtoDescriptor.builder()
+                .name(proto.getName())
+                .type(NodeType.PROTO)
+                .canonicalName(proto.getCanonicalName())
+                .filename(proto.getFilename())
+                .description(proto.getComments())
+                .build();
         write(module.getOutput() + "/data/proto/" + name + ".json", descriptor);
     }
 }
