@@ -1,5 +1,6 @@
 package io.protostuff.generator;
 
+import io.protostuff.compiler.model.ImmutableModule;
 import io.protostuff.compiler.model.Module;
 import io.protostuff.compiler.model.Proto;
 import io.protostuff.compiler.parser.ClasspathFileReader;
@@ -7,8 +8,6 @@ import io.protostuff.compiler.parser.Importer;
 import io.protostuff.compiler.parser.ProtoContext;
 import org.junit.Ignore;
 import org.junit.Test;
-
-import java.util.Collections;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -23,9 +22,10 @@ public class StCompilerTest extends AbstractCompilerTest {
         Proto proto = context.getProto();
         StCompilerFactory compilerFactory = injector.getInstance(StCompilerFactory.class);
         ProtoCompiler compiler = compilerFactory.create("io/protostuff/generator/proto/proto3.stg");
-        Module module = new Module();
-        module.setProtos(Collections.singletonList(proto));
-        module.setOutput("./");
+        Module module = ImmutableModule.builder()
+                .addProtos(proto)
+                .output("./")
+                .build();
         compiler.compile(module);
     }
 }
