@@ -3,7 +3,7 @@ package io.protostuff.generator;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 
-import java.io.OutputStream;
+import java.io.ByteArrayOutputStream;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -18,11 +18,6 @@ public class TestCompilerModule extends AbstractModule {
     @Provides
     OutputStreamFactory outputStreamFactory() {
         // dummy
-        return new OutputStreamFactory() {
-            @Override
-            public OutputStream createStream(String location) {
-                return System.out;
-            }
-        };
+        return location -> System.out;
     }
 }
