@@ -1,6 +1,8 @@
 package io.protostuff.compiler.maven;
 
+import io.protostuff.compiler.model.ImmutableModuleConfiguration;
 import io.protostuff.compiler.model.ModuleConfiguration;
+import io.protostuff.generator.CompilerModule;
 import io.protostuff.generator.ProtostuffCompiler;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -10,6 +12,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
 
 import static java.util.Collections.singletonList;
 import static org.apache.maven.plugins.annotations.LifecyclePhase.GENERATE_TEST_SOURCES;
@@ -35,14 +40,29 @@ public class JavaGeneratorMojo extends AbstractGeneratorMojo {
         super.execute();
 
         ProtostuffCompiler compiler = new ProtostuffCompiler();
-        final File sourcePath = getSourcePath();
+        final Path sourcePath = getSourcePath();
         String output = calculateOutput();
-        ModuleConfiguration moduleConfiguration = new ModuleConfiguration();
-        moduleConfiguration.setName("java");
-        moduleConfiguration.setIncludePaths(singletonList(sourcePath));
-        moduleConfiguration.setGenerator("java");
-        moduleConfiguration.setOutput(output);
-        addProtoFiles(sourcePath, moduleConfiguration);
+        ImmutableModuleConfiguration.Builder builder = ImmutableModuleConfiguration.builder()
+                .name("java")
+                .includePaths(singletonList(sourcePath))
+                .generator("java")
+                .output(output);
+        PathMatcher protoMatcher = FileSystems.getDefault().getPathMatcher("glob:**/*.proto");
+        try {
+            Files.walkFileTree(sourcePath, new SimpleFileVisitor<Path>() {
+                @Override
+                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                    if (protoMatcher.matches(file)) {
+                        String protoFile = sourcePath.relativize(file).toString();
+                        builder.addProtoFiles(normalizeProtoPath(protoFile));
+                    }
+                    return super.visitFile(file, attrs);
+                }
+            });
+        } catch (IOException e) {
+            LOGGER.error("Can not build source files list", e);
+        }
+        ModuleConfiguration moduleConfiguration = builder.build();
 
         LOGGER.debug("Module configuration = {}", moduleConfiguration);
         compiler.compile(moduleConfiguration);
