@@ -2,7 +2,11 @@ package io.protostuff.compiler.model;
 
 import com.google.common.base.MoreObjects;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -30,11 +34,9 @@ public class Enum extends AbstractUserType {
         if (constants == null) {
             return Collections.emptySet();
         }
-        Set<String> result = new LinkedHashSet<String>();
-        for (EnumConstant constant : constants) {
-            result.add(constant.getName());
-        }
-        return result;
+        return constants.stream()
+                .map(EnumConstant::getName)
+                .collect(Collectors.toSet());
     }
 
     public EnumConstant getConstant(String name) {
@@ -48,7 +50,7 @@ public class Enum extends AbstractUserType {
 
     public void addConstant(EnumConstant value) {
         if (constants == null) {
-            constants = new ArrayList<EnumConstant>();
+            constants = new ArrayList<>();
         }
         constants.add(value);
     }
