@@ -2,10 +2,11 @@ package io.protostuff.compiler.model;
 
 import com.google.common.base.MoreObjects;
 
-import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import javax.annotation.Nullable;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -72,7 +73,7 @@ public class Extension extends AbstractElement implements FieldContainer, GroupC
 
     public void addField(Field field) {
         if (fields == null) {
-            fields = new ArrayList<Field>();
+            fields = new ArrayList<>();
         }
         fields.add(field);
     }
@@ -113,7 +114,7 @@ public class Extension extends AbstractElement implements FieldContainer, GroupC
     @Override
     public void addGroup(Group group) {
         if (groups == null) {
-            groups = new ArrayList<Group>();
+            groups = new ArrayList<>();
         }
         groups.add(group);
     }
