@@ -49,7 +49,7 @@ public class Message extends AbstractUserTypeContainer
 
     public void addField(Field field) {
         if (fields == null) {
-            fields = new ArrayList<Field>();
+            fields = new ArrayList<>();
         }
         fields.add(field);
     }
@@ -86,7 +86,7 @@ public class Message extends AbstractUserTypeContainer
 
     public void addOneof(Oneof oneof) {
         if (oneofs == null) {
-            oneofs = new ArrayList<Oneof>();
+            oneofs = new ArrayList<>();
         }
         oneofs.add(oneof);
     }
@@ -129,22 +129,8 @@ public class Message extends AbstractUserTypeContainer
     }
 
     @Override
-    public String getCanonicalName() {
-        String fqn = getFullyQualifiedName();
-        if (fqn.startsWith(".")) {
-            return fqn.substring(1);
-        }
-        return fqn;
-    }
-
-    @Override
     public void setFullyQualifiedName(String fullyQualifiedName) {
         this.fullyQualifiedName = fullyQualifiedName;
-    }
-
-    @Override
-    public boolean isNested() {
-        return getParent().getDescriptorType() != DescriptorType.PROTO;
     }
 
     @Override
@@ -189,7 +175,7 @@ public class Message extends AbstractUserTypeContainer
     @Override
     public void addGroup(Group group) {
         if (groups == null) {
-            groups = new ArrayList<Group>();
+            groups = new ArrayList<>();
         }
         groups.add(group);
     }
@@ -207,7 +193,7 @@ public class Message extends AbstractUserTypeContainer
 
     public void addExtensionRange(Range range) {
         if (extensionRanges == null) {
-            extensionRanges = new ArrayList<Range>();
+            extensionRanges = new ArrayList<>();
         }
         extensionRanges.add(range);
     }
@@ -225,7 +211,7 @@ public class Message extends AbstractUserTypeContainer
 
     public void addReservedFieldRange(Range range) {
         if (reservedFieldRanges == null) {
-            reservedFieldRanges = new ArrayList<Range>();
+            reservedFieldRanges = new ArrayList<>();
         }
         reservedFieldRanges.add(range);
     }
@@ -243,7 +229,7 @@ public class Message extends AbstractUserTypeContainer
 
     public void addReservedFieldName(String name) {
         if (reservedFieldNames == null) {
-            reservedFieldNames = new ArrayList<String>();
+            reservedFieldNames = new ArrayList<>();
         }
         reservedFieldNames.add(name);
     }
@@ -259,15 +245,5 @@ public class Message extends AbstractUserTypeContainer
     public boolean isMapEntry() {
         DynamicMessage.Value value = this.getOptions().get(MessageParseListener.OPTION_MAP_ENTRY);
         return value != null && value.isBooleanType() && value.getBoolean();
-    }
-
-    @Override
-    public Enum getEnum(String name) {
-        for (Enum anEnum : getEnums()) {
-            if (name.equals(anEnum.getName())) {
-                return anEnum;
-            }
-        }
-        return null;
     }
 }
