@@ -1,11 +1,13 @@
 package io.protostuff.compiler.model;
 
 import com.google.common.base.MoreObjects;
-import io.protostuff.compiler.parser.ProtoContext;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import io.protostuff.compiler.parser.ProtoContext;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -104,18 +106,15 @@ public class Proto extends AbstractUserTypeContainer implements UserTypeContaine
     }
 
     public List<Import> getPublicImports() {
-        List<Import> result = new ArrayList<Import>();
-        for (Import anImport : getImports()) {
-            if (anImport.isPublic()) {
-                result.add(anImport);
-            }
-        }
-        return result;
+        return getImports()
+                .stream()
+                .filter(Import::isPublic)
+                .collect(Collectors.toList());
     }
 
     public void addImport(Import anImport) {
         if (imports == null) {
-            imports = new ArrayList<Import>();
+            imports = new ArrayList<>();
         }
         imports.add(anImport);
     }
@@ -133,19 +132,9 @@ public class Proto extends AbstractUserTypeContainer implements UserTypeContaine
 
     public void addService(Service service) {
         if (services == null) {
-            services = new ArrayList<Service>();
+            services = new ArrayList<>();
         }
         services.add(service);
-    }
-
-    @Override
-    public Message getMessage(String name) {
-        for (Message message : getMessages()) {
-            if (name.equals(message.getName())) {
-                return message;
-            }
-        }
-        return null;
     }
 
     @Override
