@@ -28,6 +28,8 @@ public interface UserType extends Descriptor, FieldType {
     /**
      * Test if this type is nested (declared inside of other message).
      */
-    boolean isNested();
+    default boolean isNested() {
+        return getParent().getDescriptorType() != DescriptorType.PROTO;
+    }
 
 }
