@@ -17,12 +17,8 @@ public class ProtoTreeWalker {
                      MessageOperation messageOperation,
                      EnumOperation enumOperation) {
         protoOperation.process(proto);
-        for (Enum anEnum : proto.getEnums()) {
-            enumOperation.process(anEnum);
-        }
-        for (Message message : proto.getMessages()) {
-            messageOperation.process(message);
-        }
+        proto.getEnums().forEach(enumOperation::process);
+        proto.getMessages().forEach(messageOperation::process);
         for (Message message : proto.getMessages()) {
             walk(message, messageOperation, enumOperation);
         }
@@ -31,12 +27,9 @@ public class ProtoTreeWalker {
     public void walk(UserTypeContainer container,
                      MessageOperation messageOperation,
                      EnumOperation enumOperation) {
-        for (Message message : container.getMessages()) {
-            messageOperation.process(message);
-        }
-        for (Enum anEnum : container.getEnums()) {
-            enumOperation.process(anEnum);
-        }
+
+        container.getMessages().forEach(messageOperation::process);
+        container.getEnums().forEach(enumOperation::process);
         for (Message message : container.getMessages()) {
             walk(message, messageOperation, enumOperation);
         }
