@@ -5,14 +5,15 @@ import org.antlr.v4.runtime.CharStream;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.RandomAccessFile;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+
+import javax.annotation.Nullable;
 
 /**
  * Load file from local filesystem.
@@ -23,13 +24,13 @@ public class LocalFileReader implements FileReader {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(LocalFileReader.class);
 
-    private final List<File> pathList;
+    private final List<Path> pathList;
 
-    public LocalFileReader(File... paths) {
+    public LocalFileReader(Path... paths) {
         this.pathList = checkDirectories(Arrays.asList(paths));
     }
 
-    public LocalFileReader(List<File> paths) {
+    public LocalFileReader(List<Path> paths) {
         this.pathList = checkDirectories(paths);
     }
 
@@ -37,12 +38,12 @@ public class LocalFileReader implements FileReader {
      * Check that all elements in the list exist and are directories.
      * Log warning for each element that is not directory.
      */
-    private List<File> checkDirectories(List<File> pathList) {
-        List<File> result = new ArrayList<File>();
-        for (File path : pathList) {
-            if (!path.exists()) {
+    private List<Path> checkDirectories(List<Path> pathList) {
+        List<Path> result = new ArrayList<>();
+        for (Path path : pathList) {
+            if (!Files.exists(path)) {
                 LOGGER.warn("'{}' does not exist", path);
-            } else if (!path.isDirectory()) {
+            } else if (!Files.isDirectory(path)) {
                 LOGGER.warn("'{}' is not directory", path);
             }
             // todo: should we use not existing paths? this behavior is copied from 'protoc' - it just shows warning
@@ -54,26 +55,18 @@ public class LocalFileReader implements FileReader {
     @Nullable
     @Override
     public CharStream read(String name) {
-        for (File prefix : pathList) {
-            try {
-                String pathString = resolveRelativeFile(prefix, name);
-                RandomAccessFile path = new RandomAccessFile(pathString, "r");
-                byte[] bytes = new byte[(int) path.length()];
-                path.readFully(bytes);
-                path.close();
-                String result = new String(bytes, "UTF-8");
-                return new ANTLRInputStream(result);
-            } catch (FileNotFoundException e) {
-                // ignore
-            } catch (IOException e) {
-                LOGGER.debug("Could not read/resolve {} + {}", prefix, name);
+        for (Path prefix : pathList) {
+            Path path = prefix.resolve(name);
+            if (Files.isRegularFile(path)) {
+                try {
+                    byte[] bytes = Files.readAllBytes(path);
+                    String result = new String(bytes, StandardCharsets.UTF_8);
+                    return new ANTLRInputStream(result);
+                } catch (IOException e) {
+                    LOGGER.debug("Could not read {}", path);
+                }
             }
-
         }
         return null;
-    }
-
-    public static String resolveRelativeFile(File parent, String file) throws IOException {
-        return parent.getCanonicalPath() + "/" + file;
     }
 }
