@@ -1,17 +1,28 @@
 package io.protostuff.compiler.parser;
 
-import io.protostuff.compiler.model.*;
-import io.protostuff.compiler.model.Enum;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Provider;
 import java.util.Deque;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
+
+import javax.inject.Inject;
+import javax.inject.Named;
+import javax.inject.Provider;
+
+import io.protostuff.compiler.model.Descriptor;
+import io.protostuff.compiler.model.DescriptorType;
+import io.protostuff.compiler.model.DynamicMessage;
+import io.protostuff.compiler.model.Element;
+import io.protostuff.compiler.model.Enum;
+import io.protostuff.compiler.model.Field;
+import io.protostuff.compiler.model.FieldType;
+import io.protostuff.compiler.model.Message;
+import io.protostuff.compiler.model.ProtobufConstants;
+import io.protostuff.compiler.model.ScalarFieldType;
+import io.protostuff.compiler.model.UserTypeContainer;
 
 import static io.protostuff.compiler.parser.DefaultDescriptorProtoProvider.DESCRIPTOR_PROTO;
 import static io.protostuff.compiler.parser.TypeResolverPostProcessor.createScopeLookupList;
@@ -34,18 +45,8 @@ public class OptionsPostProcessor implements ProtoContextPostProcessor {
     @Override
     public void process(ProtoContext context) {
         ProtoWalker.newInstance(context)
-                .onProto(new ProtoWalker.Processor<Proto>() {
-                    @Override
-                    public void run(ProtoContext context, Proto proto) {
-                        processOptions(context, proto);
-                    }
-                })
-                .onMessage(new ProtoWalker.Processor<Message>() {
-                    @Override
-                    public void run(ProtoContext context, Message message) {
-                        processOptions(context, message);
-                    }
-                })
+                .onProto(this::processOptions)
+                .onMessage(this::processOptions)
                 .walk();
     }
 
@@ -65,7 +66,7 @@ public class OptionsPostProcessor implements ProtoContextPostProcessor {
     private void processOptions(ProtoContext context, Message sourceMessage, Descriptor owningDescriptor, DynamicMessage options) {
         ExtensionRegistry extensionRegistry = context.getExtensionRegistry();
         Map<String, Field> extensionFields = extensionRegistry.getExtensionFields(sourceMessage);
-        Map<DynamicMessage.Key, String> fullyQualifiedNames = new HashMap<DynamicMessage.Key, String>();
+        Map<DynamicMessage.Key, String> fullyQualifiedNames = new HashMap<>();
         for (Map.Entry<DynamicMessage.Key, DynamicMessage.Value> entry : options.getFields()) {
             DynamicMessage.Key key = entry.getKey();
             DynamicMessage.Value value = entry.getValue();
