@@ -1,13 +1,20 @@
 package io.protostuff.compiler.parser;
 
 import io.protostuff.compiler.model.Element;
-import io.protostuff.compiler.model.Proto;
-import io.protostuff.compiler.model.Type;
+import io.protostuff.compiler.model.UserType;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Deque;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import io.protostuff.compiler.model.Proto;
+import io.protostuff.compiler.model.Type;
 
 /**
  * @author Kostiantyn Shchepanovskyi
@@ -29,10 +36,10 @@ public class ProtoContext {
 
     public ProtoContext(String filename) {
         this.filename = filename;
-        symbolTable = new HashMap<String, Type>();
-        declarationStack = new ArrayDeque<Object>();
-        imports = new ArrayList<ProtoContext>();
-        publicImports = new ArrayList<ProtoContext>();
+        symbolTable = new HashMap<>();
+        declarationStack = new ArrayDeque<>();
+        imports = new ArrayList<>();
+        publicImports = new ArrayList<>();
         proto = new Proto();
         proto.setContext(this);
         proto.setFilename(filename);
